driverflow/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy>=1.26,<3
pillow==10.3.0
pyngrok==7.1.6
supervision>=0.21.0
transformers==4.38.2
torch>=2.0.0
torchvision>=0.15.0

